Move accountInof to root directory
