+++ src/components/screens/transactionDetails/accountInfo.js
@@ -1,10 +1,10 @@
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import React from 'react';
-import AccountVisual from '../../../toolbox/accountVisual';
-import routes from '../../../../constants/routes';
-import styles from './transactionDetailView.css';
-import { validateAddress } from '../../../../utils/validators';
+import AccountVisual from '../../toolbox/accountVisual';
+import routes from '../../../constants/routes';
+import styles from './transactions.css';
+import { validateAddress } from '../../../utils/validators';
 
 const AccountInfo = ({
   address,
